server/pyproject.toml
[tool.poetry]
name = "innovatesg"
version = "0.1.0"
description = "InnoVateSG server for business innovation assessment"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
langchain = "^0.0.267"  # Updated from langchain-openai
openai = "^0.28.0"
python-dotenv = "^1.0.0"
# asyncio is part of the standard library for Python 3.7+, no need to specify it

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"